requirements.txt
Django==2.0.5
djangorestframework==3.8.2
mysqlclient==1.3.12
djangorestframework-camel-case===0.2.0
django-cors-headers==2.2.0
django-rest-swagger
django-model-utils===3.1.2
drf-nested-routers==0.90.0
django-environ
django-cors-headers
